sharing fixes

+++ src/backend.cr
@@ -86,23 +86,15 @@ class CrystalDrive::Backend
         STORE.unlink path
     end
 
-    def self.file_stats(path : String)
+    def self.file_stats(orig_path : String, path : String)
         file_meta = STORE.file_stats path: path
         file_meta = file_meta.not_nil!
-
-        parts = Path.new(path).parts
-        parts.delete_at(1)
-
-        if parts.size > 1 && parts[1] == @@shared_with_me_dirname
-            parts[1] = "shared"
-        end
-
-        path = Path.new parts
+        
         item = CrystalDrive::Item.new
         item.key = file_meta.not_nil!.id.not_nil!.to_s
         item.name = file_meta.not_nil!.name.not_nil!
         item.size = file_meta.not_nil!.size
-        item.path = path.to_s
+        item.path = orig_path
         item.extension = File.extname(item.name)
         item.modified = Time.unix(file_meta.not_nil!.last_modified).to_s("%Y-%m-%dT%H:%M:%S")
         item.mode = 493 #file_meta.not_nil!.mode.to_u16
@@ -196,30 +188,58 @@ class CrystalDrive::Backend
         self.is_file(path) ? STORE.file_stats(path) : STORE.dir_stats(path)
     end
 
-    def self.share(path : String, current_user : String, shares : Hash(String, String))
-        newly_added_users = [] of String
+    def self.share(path : String, current_user : String, shares : Hash(String, Hash(String, String)))
         updated_users = [] of String
         deleted_users = [] of String
 
         begin
             share = CrystalDrive::Share.get(STORE.db, path)
         rescue CrystalDrive::UserNotFoundError
-            share = CrystalDrive::Share.new path: path, permissions: Hash(String, String).new
+            share = CrystalDrive::Share.new path: path, permissions: Hash(String, Hash(String, String)).new
         end
+        shares.each do |name, v|
+            share_by_user_permission = ""
+            share_by_link_permission = ""
 
-        shares.each do |name, perm|
-            # delete
-            if perm == "" && share.permissions.has_key?(name)
-                share.permissions.delete(name)
+            user_share = false
+            link_share = false
+
+            if v.has_key?("user")
+                share_by_user_permission = v["user"]
+                user_share = true
+            end
+
+            if v.has_key?("link")
+                share_by_link_permission = v["link"]
+                link_share = true
+            end
+            delete = false
+            if user_share && share.permissions.has_key?(name) && share_by_user_permission == ""
+                share.permissions[name].delete("user")
+                delete = true
+            end
+            
+            if link_share && share.permissions.has_key?(name) && share_by_link_permission == ""
+                share.permissions[name].delete("link")
+                delete = true
+            end
+            # deleted
+            if share.permissions.has_key?(name) && !share.permissions[name].has_key?("user") && !share.permissions[name].has_key?("link")
                 deleted_users << name
-            # updated
-            elsif share.permissions.has_key?(name)
-                updated_users << name
-                share.permissions[name] = perm
-            # new
+                share.permissions.delete(name)
+            # updated / newly created
             else
-                newly_added_users << name
-                share.permissions[name] = perm
+                if !delete
+                    if !share.permissions.has_key?(name)
+                        share.permissions[name] = Hash(String, String).new
+                    end
+                    updated_users << name
+                    if user_share
+                        share.permissions[name]["user"] = v["user"]
+                    elsif link_share
+                        share.permissions[name]["link"] = v["link"]
+                    end
+                end
             end
         end
         
@@ -230,29 +250,18 @@ class CrystalDrive::Backend
         end
 
         basename = Path.new(path).basename
-
-        newly_added_users.each do |user|
-            begin
-                STORE.dir_create("/#{user}/#{@@shared_with_me_dirname}/#{current_user}", 755)
-            rescue CrystalStore::FileExistsError
-            end
-            STORE.symlink src: path, dest: "/#{user}/#{@@shared_with_me_dirname}/#{current_user}/#{basename}"
-        end
-
-        # work around, when user delete shared items
-        # share object still exist and not deleted
-        # if same folder shared again with them then we need to
-        # do this
         updated_users.each do |user|
             begin
                 STORE.dir_create("/#{user}/#{@@shared_with_me_dirname}/#{current_user}", 755)
             rescue CrystalStore::FileExistsError
             end
+            # try to add the symlink if not exists
             begin
                 STORE.symlink src: path, dest: "/#{user}/#{@@shared_with_me_dirname}/#{current_user}/#{basename}"
             rescue CrystalStore::FileExistsError
             end
         end
+
         deleted_users.each do |user|
             begin
                 STORE.unlink "/#{user}/#{@@shared_with_me_dirname}/#{current_user}/#{basename}"
@@ -266,13 +275,30 @@ class CrystalDrive::Backend
         CrystalDrive::Share.get(STORE.db, path)
     end
 
-    def self.share_delete(path : String, current_user : String)
+    def self.share_delete(path : String, current_user : String, type : String)
         basename = Path.new(path).basename
         share = CrystalDrive::Share.get(STORE.db, path)
+        deleted_users = [] of String
         share.permissions.each do |user, perm|
-            STORE.unlink "/#{user}/#{@@shared_with_me_dirname}/#{current_user}/#{basename}"
+            if share.permissions[user].has_key?(type)
+                share.permissions[user].delete(type)
+            end
+
+            if !share.permissions[user].has_key?("user") && !share.permissions[user].has_key?("link")
+                deleted_users << user
+                STORE.unlink "/#{user}/#{@@shared_with_me_dirname}/#{current_user}/#{basename}"
+            end
         end
-        CrystalDrive::Share.delete(STORE.db, path)
+
+        deleted_users.each do |u|
+            share.permissions.delete(u)
+        end
+
+        if share.permissions.size == 0
+            CrystalDrive::Share.delete(STORE.db, path)
+        else
+            share.save(STORE.db)
+        end
     end
 
     def self.share_link_get(path : String, permission : String, owner : String)
@@ -285,18 +311,18 @@ class CrystalDrive::Backend
 
     def self.share_link_create(hash : String, current_user : String)
         share_info = CrystalDrive::ShareLink.get(STORE.db, hash)
-        perm = {current_user =>  share_info["permission"]}
-        # if user has no previous share, create
-        # if user has previous share with less privileges, update
         
-        # self.share(share_info["path"], share_info["owner"], perm)
+        if share_info.size == 0
+            raise CrystalDrive::NotFoundError.new
+        end
+        perm = {current_user =>  {"link" => share_info["permission"]} } 
+        self.share(share_info["path"], share_info["owner"], perm)
         share_info
     end
 
     def self.share_link_delete(path : String, permission : String, current_user : String)
-        # ddelete only if the only share user has is link
-        # self.share_delete(path, current_user)
         CrystalDrive::ShareLink.delete(STORE.db, path, permission)
+        self.share_delete(path, current_user, "link")
     end
 
     def self.user_add(username : String, email : String)
